% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{Shannon diversity function}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector of numeric values representing
the species abundance in a community}
}
\value{
Shannon diversity value
}
\description{
Shannon diversity function
}
\examples{
com <- c(1,4, 6, 8)
my_shannon(com)

}
